Match effect entry count width
Carry the outer effect entry count as an unsigned word while retaining its halfword load, eliminating the final count-normalization mismatch and reaching the target function size.

Constraint: Preserve the recovered entry initialization and frame reload lifetimes.
Confidence: high
Scope-risk: narrow
Directive: Keep the outer count unsigned; remaining residuals are register coloring and constant-pool labels rather than missing logic.
Tested: exact effect_visual_r50_80139AC4_prefix build; focused objdiff 98.333336% to 99.081635% at exact 588-byte size; git diff --check.
Not-tested: runtime particle behavior.

+++ src/game/effect/effect_visual.c
@@ -1435,7 +1435,7 @@ BOOL fn_80139AC4(void* ptr, u32 tick) {
     u32 frame;
     u16 endFrame;
     u16 i;
-    u16 count;
+    u32 count;
 
     if (ptr != NULL) {
         p = ptr;
